app/requirements.txt
fastapi
uvicorn
pydantic
azure-identity
azure-digitaltwins-core
python-dotenv
pydantic-settings
azure-cosmos
python-multipart
pandas 
numpy
azure-identity
dotenv
xgboost
scikit-learn
joblib
openpyxl
xlrd
pandas-stubs
python-dateutil
matplotlib
seaborn
plotly
statsmodels
requests
pytz
scipy
matplotlib-inline
python-multipart
